Crate acpi_table_handler

Handles ACPI tables based on signatures.

Essentially a giant multiplexer that invokes the proper table-specific handler function based on an ACPI table signature.

Functions

acpi_table_handler

The single arbiter of ACPI Table types, which contains a large table of ACPI signatures to specific table handler functions. Each handler invoked by this function will add the specific table to the given list of AcpiTables.