Function acpi_table_handler::acpi_table_handler

pub fn acpi_table_handler(
    acpi_tables: &mut AcpiTables,
    signature: AcpiSignature,
    length: usize,
    phys_addr: PhysicalAddress
) -> Result<(), &'static str>

The single arbiter of ACPI Table types, which contains a large table of ACPI signatures to specific table handler functions. Each handler invoked by this function will add the specific table to the given list of AcpiTables.

Arguments

acpi_tables: a mutable reference to the AcpiTables that will contain the new table.

signature: the signature of the ACPI table that is being added. This determines which handler is invoked.

length: the total length of the table, which was obtained from its Sdt header when originally mapped.

phys_addr: the PhysicalAddress of the table, which is used for determining where it exists within the mapped region.